feat(book): implement proper exceptions in BookController

+++ src/main/java/com/springpractice/bookstore/controller/BookController.java
@@ -1,7 +1,9 @@
 package com.springpractice.bookstore.controller;
 
+import com.springpractice.bookstore.exceptions.ResourceNotFoundException;
 import com.springpractice.bookstore.model.Book;
 import com.springpractice.bookstore.repository.BookRepository;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,15 +24,14 @@ public class BookController {
     }
 
     @PostMapping
-    public Book createBook(@RequestBody Book book) {
+    public Book createBook(@Valid @RequestBody Book book) {
         return bookRepository.save(book);
     }
 
     @PutMapping("/{id}")
-    public Book updateBook(@PathVariable long id, @RequestBody Book book) {
-        Book existingBook = bookRepository.findById(id).orElse(null);
-        if (existingBook == null) {
-            return null;
+    public Book updateBook(@PathVariable long id, @Valid @RequestBody Book book) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book with id " + id + " not found");
         }
         book.setId(id);
         return bookRepository.save(book);
@@ -39,11 +40,16 @@ public class BookController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteBook(@PathVariable long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book with id " + id + " not found");
+        }
         bookRepository.deleteById(id);
     }
 
     @GetMapping("/isbn/{isbn}")
     public Book getBooksByIsbn(@PathVariable String isbn) {
-        return bookRepository.findByIsbn(isbn).orElse(null);
+        return bookRepository.findByIsbn(isbn).orElseThrow(
+                () -> new ResourceNotFoundException("Book with ISBN (" + isbn + ") not found")
+        );
     }
 }
